requirements.txt
supabase==2.3.0
python-dotenv==1.0.0
google-cloud-vision==3.5.0
pdf2image==1.17.0
Pillow==10.2.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
flask==2.3.3
requests==2.31.0
PyMuPDF==1.23.7
openai==1.3.5
